feat(event date component): show the Time when an Event was Created
create component that can be used to display the time an event was created with it's test
