+++ src/components/__tests__/EventDate.test.jsx
@@ -1,0 +1,8 @@
+import renderer from "react-test-renderer";
+
+import EventDate from "@/components/events/EventCard/EventDate";
+
+it("renders correctly", () => {
+    const tree = renderer.create(<EventDate />).toJSON();
+    expect(tree).toMatchSnapshot();
+});
